Move all save file related objects to save-file/
