+++ src/app/models/index.ts
@@ -1,8 +1,9 @@
-export * from './skill';
-export * from './server';
-export * from './character';
-export * from './condition';
-export * from './character-name';
-export * from './inventory-slot';
 export * from './item-data';
-export * from './character-customization';
+
+export * from './save-file/skill';
+export * from './save-file/server';
+export * from './save-file/character';
+export * from './save-file/condition';
+export * from './save-file/character-name';
+export * from './save-file/inventory-slot';
+export * from './save-file/character-customization';
