+++ src/app/models/save-file/condition.ts
