+++ src/app/models/save-file/inventory-slot.ts
